+++ cmd/config.go
@@ -22,6 +22,7 @@ var (
 
 func defaultConfig() {
 	for key, val := range map[string]interface{}{
+		"update.automatic.enabled": false,
 		"update.check.enabled":     true,
 		"update.check.period":      time.Duration(time.Minute * 10).String(),
 	} {